cafesys/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.
django-rosetta
#python-dateutil # included in base.txt
reportlab
pyPdf
python-ldap
#Geraldo
numpy
django-sentry
